# Performance Optimizations and Benchmarking

## Summary
- Refactored core calculation functions for Greeks, implied volatility, and pricing.
- Added and updated micro-benchmarks to track performance regressions.
- Updated project metadata and dependencies.

## Changes
- **Cargo.toml**: Updated project metadata and added benchmarking dependencies.
- **src/greeks.rs**: Optimized implementations of `delta`, `gamma`, `theta`, `vega`, `rho`, and combined `all_greeks`.
- **src/implied_volatility.rs**: Enhanced `calc_rational_iv` convergence and reduced execution time.
- **src/pricing.rs**: Streamlined `calc_price` for lower overhead.
- **src/lib.rs**: Re-exported optimized modules and adjusted public API.
- **benches/greeks.rs**: Added a new benchmark suite to measure performance of Greeks calculations.
- **terminal.logs**: Captured benchmark outputs to document baseline and improved timings.

## Benchmark Results

### Greeks
| Metric     | Before              | After               | Change         |
|------------|---------------------|---------------------|----------------|
| Delta      | 35.99 ns            | 16.06 ns            | –55.4%         |
| Gamma      | 16.20 ns            | 0.88 ns             | –94.6%         |
| Theta      | 52.14 ns            | 18.94 ns            | –63.7%         |
| Vega       | 16.19 ns            | 0.86 ns             | –94.7%         |
| Rho        | 35.74 ns            | 17.49 ns            | –51.1%         |
| All Greeks | 1.22 µs             | 112.6 ns            | –90.8%         |

### Implied Volatility
- **calc_rational_iv**: 373.8 ns → 359.9 ns (–3.6%)

### Pricing
- **calc_price**: 31.42 ns → 30.87 ns (–1.8%)

## Motivation
Improving computational performance for high-throughput financial analytics and ensuring low-latency pricing and risk metrics.